IOS-1943 Refactoring: extracted code to save/load UserWallets
